Separate config/base/auctions to own files
