+++ alerts/config/envloader.go
@@ -6,6 +6,9 @@ import "os"
 type EnvLoader struct {
 }
 
+// Verify interface compliance at compile time
+var _ ConfigLoader = (*EnvLoader)(nil)
+
 // Get retrieves key from environment
 func (l *EnvLoader) Get(key string) string {
 	return os.Getenv(key)
